pin changes done according to actual connection

+++ include/ir_sensor.h
@@ -4,6 +4,6 @@
 #include <Arduino.h>
 
 void sensorInit();
-int readLine();
+float readLine();   // ✅ ONLY ONE declaration
 
 #endif
